removed deprecated helper function
